Fixed engine drawing code to draw Blender properties from blend file.
When opening a blend file of the same version, the engine/runner
parameter properties should be retained normally in the collection
property named "plug_val_list". These are the dynamically created
properties that are re-loaded whenever a plug changes.

However, previous versions of the code did not display these properties
because the "ordering index" key ('_i') isn't stored as a property. So
it needed to be initialized in the loading code. The quick fix is to
also initialize the ordering index in the drawing code. That's what this
commit does.

Note that this fix will preserve the engine/runner settings as Blender
properties between runs as long as the CellBlender ID isn't changed (in
other words, the same version of CellBlender is used. It does not
preserve the engine/runner settings when an upgrade is required. That
will require adding these values to the Data Model ... which is the next
thing to be done.