Split apart "-seed" and "#" (space separated) in command line runner.
For some reason, issuing a Popen command for the command line runner
with a seed specified as a string like "-seed 1" wasn't being recognized
by MCell. It was issuing an error:

   ...user_path/build/mcell: unrecognized option '-seed 1'
   MCell 3.3 (commit: 7030a99  date: Thu, 19 May 2016 11:14:16 -0400)

In other words, the Popen call was passing the entire unbroken string
as one of the arguments. This hasn't been a problem for the other
runners, but then they use an intermediary program (either another
executable like SimControl or the Java interpreter). It's possible that
they were further splitting arguments by spaces before passing to MCell.

This change splits any arguments by spaces when mapping from the command
line engine's parameter list into the actual list passed to Popen. It
may be good to watch for similar problems with the other runners.