Moved run control to Output/Control Options row to prepare for merge
Added a few minor changes to make this version more compatible with the
development branch. In particular, the default run control has been set
back to "MCell via Queue Runner" which should be the default for most
people using the development branch. But the run control selector is now
visible without opening the Output/Control Options panel to make it
easier to choose the Engine/Runner option.

The former "Dynamic" option has been changed to "Engine/Runner" to more
properly reflect its functionality.

With these changes (and possibly some testing) this might be a good time
to merge the "integrate_dynamic_runners" branch into the "development"
branch. There is still more work to be done on the dynamic runners
themselves, but that needed work will hopefully (again with some
testing) not cause problems for anyone not using the dynamic feature.