Added code to handle relative or absolute paths in MCellR Engine
This code should (hopefully) follow the convention that any path
starting with os.path.sep ("/" on *nix) is an absolute path. Any
other path will have the shared path prepended to it.

The shared path itself is handled in a similar manner. If it starts
with os.path.sep, then it is considered absolute. If not, then it is
considered relative to the cellblender/extensions directory.

All paths are eventually resolved to absolute paths before use.