Queue Runner: Created new texts directly rather than via an operator
Previous code used an operator to create new texts which did not return
a handle to the new text. Then the last text was assumed to be the newly
created one. Unfortunately, Blender was sorting them by name, so the
last text might not have been the one just created (easily proved with a
text named "zzz.txt"). In that case, the Queue Runner would begin
appending its output to the end of the file whose name sorted to the
last position in the texts list.

This was not easily noticed because the text logs started with "t" and
were already sequential ... so it worked fine.

This fix calls bpy.data.texts.new ( name ) in place of the operator.