Added a new "blender_opengl" runner to draw results in 3D View window
This is currently a copy of the command line runner since it's the
simplest runner to start with. The goal is to display the run output
directly onto the 3D view window in a scrolling manner. The runner might
include options like alpha channel for text as well as abbreviated, row
based or column based output. This is still highly experimental.