Added an OpenGL text "overlay" display to the Queue Runner
This commit adds a few buttons to the Queue Runner to control a text
"overlay" (text drawn with a 0.5 alpha channel). The text overlay can be
shown or hidden, and it can be scrolled up and down via operators.

This version only shows the last text buffer, but it shouldn't be
difficult to add a selector. One of the nice features would be sharing
the same scroll position when switching between runs (if desired). This
would be helpful when looking for places where runs might differ
(assuming that there is something more useful than iteration number
being drawn).

There are a number of problems with this current version, but it is good
for evaluating whether this kinds of "Heads Up Display" would be useful.
One easy improvement would allow color selection for the text overlay.