User-settable timer value and status messages.
Added a "Timer" value to the Dynamic Queue Runner to select the update
time. The previous value (secs) had been set to 0.5 which was fine for
the status message, but was somewhat jerky when scrolling the output
results into the 3D View overlay. This change allows the user to set
that value. A value of 0.1 is much smoother but may impact other areas.

Also added some print statements to diagnose the delay when the
simulation pauses with "Need to run DYNAMIC". It appears that the delay
comes from deleting the data from the previous run (and possibly
recreating the directory structure for the new run). This delay may be
more pronounced on systems with slower storage devices (networked).