Removed the blender_opengl runner, functionality now in Queue Runner
The blender_opengl runner was intended to provide a text overlay
display. However, that functionality was more easily added to the Queue
Runner, so the blender_opengl runner is no longer needed. This commit
removes those files and removes them from cellblender_source_info.py.

This is an appropriate time to merge the blender_opengl_draw branch back
into the development branch.