Store runs in dictionary referenced by PID for text overlay display.
This version builds a global dictionary named "screen_display_lines"
which has an entry for each PID. The value for each PID is the list of
lines that should be displayed for that PID.

Individual PIDs can be selected from the run list to display its results
in the text overlay.

The position (scroll) settings are retained across texts so it's easy to
compare certain parts of runs against one another as long as they are of
nearly identical length (typically the case for multiple run scenarios).