Moved functions into object for python engine and java runner - Works.
Right now, the Limited Python engine and the Java runner can both be run
as classes (new system) or as modules (old system). The same functions
have been provided either at the module level or inside a class.
Eventually, the module-level functions will be removed.

Note that the module level parameter_dictionary and parameter_layout can
remain at the module level. These will provide a static interface to the
Blender property system. When a new engine or runner object is created,
it will get a copy of those structures at the time of their creation.
This gives some degree of persistence to those values while the run is
taking place.