Moved obj functionality into original engines/runners; removed obj vers
It turns out that the object-based versions and the module-based
versions can coexist in the same file since the class (for the object)
separates the namespaces of the functions. In fact, the same exact
functions might be able to be used in both places. However, this could
be confusing, so purely object-based versions will be the final goal.

The temporary engine "limited_python_obj" and the temporary runner
"java_object" have been removed in this commit.