Converted the Smoldyn248 engine to the new class-based design
Note that the postprocess function presented a challenge because the
current functions (operators) are not called with a "self". So it wasn't
clear how to pass the proper engine object to the postprocess function.

The quick (and dirty) solution was to maintain a global variable named
"last_engine_object". When an object version is run, then that variable
will be set to the object itself (where the data can be found). When the
method version is run, that variable is set to "None" indicating that
the global values should be used (as before).

The better solution would be to always pass the "self" of the class to
the operator functions so they will always know what data to work on.
Hopefully that can be done in the future.