Added plug_complete to all engines/runners to control display in panel
The engine/runner interface has contained a plug_active flag, but it
wasn't widely used. This commit adds a "plug_complete" flag intended to
indicate whether or not a particular plug is ready for use by end users.
This flag is gated by the existing "developer_mode" flag such that plugs
rated at or above 0.5 will always be shown, while plugs rated less than
0.5 will only be shown in developer mode.

Note that this is a somewhat crude way to implement this functionality.
It might be better to be able to enable or disable individual plugs via
the user interface, but that functionality isn't needed right now.