Used final_shared_path logic from MCellR runner
This version should try to find its apps in the extensions folder below
the cellblender folder of the addon. It will also look for BNG2.pl in
the bng2 folder which is inside the bionetgen folder. The default final
path should look something like this:

  ..addons/cellblender/extensions/bionetgen/bng2/BNG2.pl

This default behavior is defined by using a "final_shared_path" of:

  ..addons/cellblender/extensions

and a BioNetGen path of:

  bionetgen/bng2/BNG2.pl

These will be joined to form the default mentioned earlier.