Added a Blender Data Model Browser in the Scripting Panel
This commit adds a new expandable panel inside the scripting panel to
display the full data model. This is similar to the "Browse Data Model
with Tk" function except that the data model is displayed as Blender
properties in the Data Model Browser panel.

The data model properties are sorted by key and nested properly for both
dictionaries and lists. The "From Data Model" button honors the settings
for including geometry and scripts (Data Model Scripting / Make Script).

The Data Model Browser can also create the same tree for any dictionary
in a Blender internal text by setting the "Internal Text Name" and using
the "From File" button. The "To File" button is not functioning at this
time. The planned functionality would allow for changing values and
restoring them either to a Blender internal text file or regenerating
the Blender properties from the modified data model.

At this time, the data model is displayed fully expanded, and only the
top level show/hide button works.