Added open/close all operators and show "name" fields for dictionaries
The open/close all operators fully expand and collapse the tree.

If any dictionaries have a "name" field, that field will be shown at the
level of the dictionary itself (helpful for lists of objects, molecules,
and other dictionaries with a "name" field). If a plain "name" field
isn't found, the dictionary will be searched for any fields containing
"name" (such as "mol_name"). The first found (if any) will be shown.
Note that this name searching is currently case-sensitive for lower case
only. Most (likely all) CellBlender naming keys are lower case.