Temporarily increased the OpenGL runner's priority to show up by default
The OpenGL runner is good to use for testing, but it doesn't show up by
default when it's "plug_complete" value is less than 0.5 (unless the
developer mode option is set). To keep from having to do this often, the
priority (completeness) value was changed to be over 0.5.

Also, some comments have been added to clarify usage in
cellblender_simulation.py.

Also, a missing function was restored for Smoldyn module mode operation.