Changes from last night ... not sure if this even runs.
These changes were made up to 11pm last night, but have not been tested.

There may also be a better approach than having each task in the task
list correspond to one active runner. Maybe a runner should be a group
of tasks that can be managed either as a group or individually. The
current task list does not provide for that case.

For large production runs, it's easy to imagine hundreds of MCell tasks
being spawned for a single "run". If multiple runs are made, then there
isn't currently a mechanism to just terminate those from a single "run".

If we design a task list where the individual tasks retain an
association with the "run" that spawned them, then we can support
managing the group for each "run" together. One implementation of this
would use a template list of "tuple" properties where the first index in
the tuple is the "run" and the second is the task within that run. The
second index might also be the raw process ID (although this might be
better stored in the job itself).