Updated the README.md file to reflect the newer directory layout names.
The data_layout.json file was originally defined to use several special
keys: ("dir", "file_type", and "SEED"). When sweeping is used, those
names could conflict with legal parameter names, so a special convention
was implemented to use all caps for these "special" keys and to precede
them with a slash ("/"). This was implemented in commits:

  7894e67963df5b4ffdaddf646d26fa058bc5c52f
  af7ca5b3ce587830a3c07c78a6fdf2800fb68ef4

However, the project's README.md file had not been updated.