Added special handling for relative paths to MCell Binary in Blender
Blender supports file specifications that are relative to the ".blend"
file using a "double slash" syntax:

  //../../../../../mcell_dg/build/mcell

However, this syntax has not been supported in some areas of
CellBlender. This commit handles this syntax properly in the MCell
simulation engine. Changes to other files were made in the process
(printing, for example, when the mcell binary is found invalid).