Removed much of the module-level code.
The conversion of engines and runners to classes has been gradual. While
CellBlender understands the new classes, it will fall back to the module
level versions if the class versions are not found.

This commit removed (or commented out) the module level code for those
that have been converted to the new class system. The decision whether
to actually remove the old module code or simply comment it out was
based on whether it had been tested yet or not.

Also note that the Job Timer keeps running with the Local Queue runner
when running anything other than MCell (dyn geo or data model). It is
possible that the timer is looking for something in the MCell stream to
flag completion, but that doesn't seem to have been a likely design.
It's also not clear right now if there may be 2 timers being started
with the local queue version since it still piggy-backs off of some of
the non-dynamic Queue Runner code.