Added sweeping support to Queue Runner, disabled most others.
This commit simplifies the current run options to include only:

  MCell Local
  MCell SGE
  Engine/Runner (Experimental)

The "MCell Local" option is a combination of:
  Queue Runner (with sweeping added)
  MCell 3 Data Model (already supported sweeping)

The "MCell SGE" option is a combination of:
  SGE Runner (already supported sweeping)
  MCell 3 Data Model (already supported sweeping)

The "Engine/Runner" option contains other combinations. This option is
titled "Experimental" because it is not finalized and all options do not
work yet. In preparation for release, this option might be hidden with a
"developer's mode" flag. That's a policy decision to be made.

At this time, dynamic geometry is only supported with the Engine/Runner
option. This is because the other 2 options (MCell Local and MCell SGE)
are based on the data model and the data model does not currently
contain the computed dynamic geometry. It does, however, contain the
information needed to compute the dynamic geometry either by itself
(when scripted dynamic geometry is used) or in conjunction with the
".blend" file (when shape key dynamic geometry is used).

In order to support dynamic geometry in the two non-experimental modes,
the exporters will need to have the dynamic geometry generation code
added. This is not anticipated to be too difficult, and is the last
major proposed addition before the upcoming CellBlender release.