This is a partially working dynamic geometry version.
The original dynamic geometry exporting was done by patching up the
files exported as modular MDL files. However, since the data model
exporter exports a single unified MDL, the same "patching up" code
doesn't work on that exporter.

This commit was an attempt to fix the "patching up" code to work with
the unified MDL file, but it was trickier than expected and is not
complete (or working) in this version.

The proper solution is to do the "patching" as the original MDL files
are being written (in other words, write them as dynamic geometry MDL
files from the beginning). That should be the final step to get this
working before release.