Working dynamic geometry with sweepable parameters in "MCell Local"
This version exports all objects as dynamic if any object is dynamic.
This is required to work with the current MCell dynamic geometry
implementation.

Note that sweeping a dynamic geometry model requires considerable file
space since the dynamic geometry files are exported for each point in
the sweep space.

Note that this version currently requires one non-swept run to generate
the dynamic geometry files at the top level when displaying from "File"
(as opposed to "Script" or "Other"). Future versions may look for their
dynamic geometry files in the selected sweep directory (just as the
molecule files are handled now).