Updated for SGE sweeping and dynamic geometry. Default to MCell Queue.
Much of the non-scripted dynamic geometry exporting code had depended
directly on Blender. This code was ported to "mdl/data_model_to_mdl.py"
for the initial testing of the sweepable dynamic geometry combination.
However, this would not work on versions that expected to get all data
from a data model (the Galaxy environment, for example). It was also a
problem with the SGE runner which also expected to get its data from a
data model. This commit updated mdl/data_model_to_mdl.py to be able to
generate dynamic geometry directly from a data model without Blender.

This version has 3 hard-coded engine/runner combinations as well as the
previous Engine/Runner support for pluggable modules. This provides the
following 4 top-level choices:

  MCell via Queue Runner (default)  - Dynamic Geometry
  MCell Local                       - Dynamic Geometry and Sweeping
  MCell SGE                         - Dynamic Geometry and Sweeping
  Engine/Runner (Experimental)      - All other combinations

The first option (MCell/Queue) is the default because that's been the
best tested of all. It still uses the io_mesh_mcell_mdl exporting
mechanism and supports dynamic geometry, but not sweeping.

The second option (MCell Local) also uses the Queue, but rather than
exporting from Blender properties, it exports from a data model. This
uses the same python exporting code that was developed for Galaxy. It
has always supported sweeping, but this commit adds dynamic geometry as
well. After this has been tested, it should become the default for the
upcoming release of CellBlender (sweeping and dynamic geometry).

The third option is the Sun Grid Engine runner that has also been
updated to support dynamic geometry. It also inherits its sweeping
capabilities from the Galaxy work because it uses the same python
exporting code (mdl/data_model_to_mdl.py) that it shares with the
MCell Local option.

The fourth option is the "buffet" which still provides the ability to
mix and match all engines and runners. It provides support for CBNGL,
MCellR, and other experimental engines. It will also be the proving
ground for new runners like PBS as they are developed.

Note that the mdl/data_model_to_mdl.py code has been in use for a long
time, but it has not undergone extensive testing for handling the more
unusual cases. This should be done as we approach this upcoming release.