Force single precision output for MISSED_REACTION_THRESHOLD to pass test
Some of the test suite values had been exported as floats with %.15g
since May 18th, 2016 (commit cb69df07fff). Since Blender's internal
storage is float (not Python's default double), these values were
exported from Blender with the precision of floats.

Now that we're using the data model (which stores values as either
strings or double), the data model export writes these slightly more
accurate values to the MDL files. This is good.

However, our test suite stores a hash of the written MDL files, so this
increased accuracy in the MDL files is considered a test failure (it
doesn't generate the EXACT SAME MDL as it did before).

Since our exporting software is being re-written to handle both dynamic
geometry and parameter sweeping, it is important to verify that the new
export code generates **EXACTLY** the same output as the old export
code. This will require (at least temporarily) exporting the double
values in the data model as floats (not doubles) when writing MDL.

This commit is the first to force float exporting to MDL. In this case,
the MDL parameter "MISSED_REACTION_THRESHOLD" is forced to float on
export.

Also note that prior to commit cb69df07fff611e3e63feb5f5d7941e2bedfa6e7,
most values were exported as %g which rounded them to look like the
values entered as strings in the panel ("0.001" would be "0.001").
However, with the introduction of %.15g format in that commit, a value
like 0.001 would be exported as 0.00100000004749745. With the new double
precision, 0.001 is now exported (via %.15g) as 0.001. This is what
breaks the test suite and forces the introduction of this commit.