Fixed export to match all non-geometry and simple-geometry test cases
There have been a number of test cases that no longer get the exact same
answers with the new data model export code.

As mentioned in previous commits, this is because the path from string
(exact) to data model (python double) to MDL is of greater precision
than any path that includes a Blender float. To get the exact same
results, some of the data model values had to have their precision
reduced to that of a float before export. This commit fixed that for a
number of exported values.

Another difference appears to be the export ordering of objects. When
objects are exported from Blender's object list, they appear to be
sorted alphabetically. But they are not stored alphabetically in the
data model. It's not clear why they should be different since the data
model is built from Blender's properties, but this commit sorts the
objects by name before exporting.

This commit fixes most of the tests that were not passing (see below).
The remaining failed tests are the dynamic geometry tests and 2 of the
complete model tests. The organelle test is producing nearly identical
MDL with the exception of a missing "Scene." in the boolean geometry
expression for both Organelle's. This is likely related to the dynamic
geometry problems.

===================================================================

These test cases were failing and fixed by this commit:

  NonGeometry:     6, 7, 8, 9
  SimpleGeometry:  5, 6, 9
  CompleteModel:   1, 2

The following tests are still not passing:

  DynamicGeometry: All
  CompleteModel:   3, 4