This commit can almost run dynamic geometry via the Sun Grid Engine.
The SGE runner uses a data model along with data_model_to_mdl.py to
generate the MDL on the "remote" machine. This commit adds the dynamic
geometry to each non-scripted dynamic object via a new key named:

  frame_list

Each frame contains the same "vertex_list"/"element_connections" fields
that are stored in each object's normal (static) geometry. These frames
are only included if the object is dynamic but does not have a script to
describe the dynamic geometry.

The current version works properly for scripted dynamic geometry in both
of our major runners (MCell Local and MCell SGE). However, non-scripted
dynamic geometry (typically via shape keys) only works with MCell Local
and not with MCell SGE. The problem appears to be in the generation of
the dynamic data in the data model. Inspection shows that the static
geometry is output for every frame. So while the file structure seems to
be appropriate for dynamic geometry, the actual values are not changing.

Given that we've never had dynamic geometry running on the Grid Engine,
this should not break any existing uses of the development branch.