This commit can run dynamic geometry via the Sun Grid Engine
This commit now supports dynamic geometry and parameter sweeping for
MCell running locally (Queue runner) and remotely (SGE runner). The
previous commit did not support shape-key based dynamic geometry with
the SGE runner. This commit supports both shape key and scripted dynamic
geometry for both runners.

The method used to export dynamic geometry from shape keys is different
than the method used to export static geometry. In particular, the
"to_mesh" method must be called with apply_modifiers set to True, and
with the settings set to "PREVIEW", and with calc_tessface set to False.

This may be a more computationally expensive way to get the vertices,
and future versions might use the previous (direct mesh) method for
objects that do not have shape key dynamic geometry.

Also, using this method (at least as currently implemented) appears to
"bake" the location (offset) into the mesh coordinates. So the location
is being intentionally set to zero. It would be better if the transform
could be done without applying the object's location to the mesh to
preserve the location as an object property (as MCell supports).

Finally, this version contains a fair amount of debug printing. It
should be removed in subsequent commits.