Added code to put dynamic geometry surface regions into the data model.
The data_model_to_mdl.py program did not include surface regions in the
data model for dynamic geometry objects. This commit adds that support.

Note that there is a problem with script-generated dynamic geometry
objects that have surface regions. These surface regions are normally
added through Blender's panels. They appear to be attached to the
"mcell" key for each object, but that may not be being created for
script-generated dynamic objects. This will need more work.