Create dynamic geometry at each sweep point / Add region_props
This commit creates and uses dynamic geometry at each point in a
parameter sweep. This will be needed for sweeping dynamic geometry.
This can generate a lot of data and future versions may not export
redundant data.

Also added a "region_props" variable to the environment that runs
the dynamic geometry scripts. Dynamic geometry scripts that create
regions will need this information to export proper MDL. It isn't
being used yet in this version, but it was added in anticipation of
upcoming changes.