Add surface classes to scripted dynamic geometry - not working with MCell yet.
This commit adds a "SURFACE_CLASS = name" line inside the DEFINE_SURFACE_REGIONS
block inside a POLYGON_LIST block included in a dynamic geometry frame file.

This produces a frame file for each object that looks like this:

  Cube POLYGON_LIST
  {
    VERTEX_LIST
    {
      [ 1.0, 1.0, -1.5 ]
      [ 1.0, -1.0, -1.5 ]
      [ -1.0, -1.0, -1.5 ]
      [ -1.0, 1.0, -1.5 ]
      [ 0.1, 0.1, 1.5 ]
      [ 0.1, -0.1, 1.5 ]
      [ -0.1, -0.1, 1.5 ]
      [ -0.1, 0.1, 1.5 ]
    }
    ELEMENT_CONNECTIONS
    {
      [ 1, 2, 3 ]
      [ 7, 6, 5 ]
      [ 4, 5, 1 ]
      [ 5, 6, 2 ]
      [ 2, 6, 7 ]
      [ 0, 3, 7 ]
      [ 0, 1, 3 ]
      [ 4, 7, 5 ]
      [ 0, 4, 1 ]
      [ 1, 5, 2 ]
      [ 3, 2, 7 ]
      [ 4, 0, 7 ]
    }
    DEFINE_SURFACE_REGIONS
    {
      top
      {
        ELEMENT_LIST = [1, 7]
        SURFACE_CLASS = transp /* <<<--- MCell Error Line --- */
      }
    }
    TRANSLATE = [ 0, 0, 0 ]
  }

As shown by the comment, MCell complains about the SURFACE_CLASS line
that is included in the SECOND frame file of the dynamic geometry (it
appears to be OK in the first frame file).

It would be helpful if MCell could accept this syntax so that dynamic
geometry scripts could generate surface class assignments as they are
writing out the rest of the geometry.

If this can't be supported, then the MODIFY_SURFACE_REGIONS instruction
would have to be generated based on some cumulative result of all the
dynamic geometry scripts.

Note that this is not a problem with Shape Key based dynamic geometry
because CellBlender has an actual object that will not change its faces
during the run (only the vertices change via the shape keys). But with
scripted dynamic geometry, the script cannot know which faces of the
"original" object should map to which faces that it will generate
dynamically. The script only needs the original object as a "vessel"
to hold the new mesh data that will be added. In fact, you can start
off with any object (cube, torus, cone), and it will become whatever
the dynamic geometry script creates at frame 0. So it doesn't make
sense to try to define surface regions on the original object within
Blender.