Region support for DynGeo script, partial Export-only support
The regions_dict and regions_props dictionaries needed to
be available in the update_scene function that calls the
user's dynamic geometry script (since the script expects
those variables to exist). This commit adds them.

Current separation of export and run in sweeps could result
in duplication of the sweep code or require temporary storage
of the sweep commands for running. Alternatively, (as in this
commit), flags could be used to signal which sections of the
export/run code should be executed.

This commit adds support for the two separate flags, but the
operators have not been updated to set them yet.

Also, much of the export code is in "cellblender_project.py".
Given the tight coupling between exporting and running, this
code may more properly belong in "cellblender_simulation.py".
This is a good topic for discussion.