Added support for separate export and run with dynamic geometry & sweeps
This commit adds the ability to just export and just run a project with
both dynamic geometry and sweeping (using the "MCell Local" run option).
This finished the work of the previous commit.

This version does the sweeping in cellblender_simulation.py and creates
the dynamic geometry in data_model_to_mdl.py.

Note that there was a minor problem with this separation of exporting and
running based on CellBlender's "user protection" mechanisms. It turns
out that when exporting and running are decoupled, CellBlender will only
show the export button (and no other settings) when it can find a main
MDL file in the project's target directory (files/mcell/output_data).
The old exporter was still generating those files even during sweeps
when the actual MDL files were down in the swept directories. Since this
version no longer generates those "top level" (unused) MDL files, the
"user protection" mechanism of CellBlender would not allow running ...
even though the files had been exported. Since this checking is a bit
complicated, the quick solution was to generate a mostly empty MDL file
at the top where CellBlender is looking (it contains just one comment).
This should probably be fixed in the future.

Note that this commit has not been well tested. So if you have trouble,
back up a few commits.