Disable export of SGE models without running since it isn't supported.
The original (pre engine/runner) SGE option would export and run at the
same time. That's the version being used for this release. So exporting
would do both export and run. This behavior is inherited from the export
and run script originally coded for the Galaxy platform.

The engine/runner model implicitly separated exporting from running
because it was the engine's job to export and produce run commands, and
it was the runner's job to execute the commands and provide a user
interface for job control.

This commit refuses to "export" from an SGE job and instructs the user
to disable the "Decouple Export and Run" setting in order to run.