Tests: ..ops.mcell.export_project() --> ..ops.mcell.dm_export_mdl()
Forced the test suite to export with the new data model version.

Prior to this commit, the test suite was running with the new data model
version, but when exporting alone (not running), it was still exporting
with the old mechanism. This caused considerable confusion when testing
the new data model version without running.