Updated CellBlender and test suite to better handle dynamic geometry.
Added "large_molecular_displacement" to the data model and "Warnings"
panel. The current default value is "WARNING" since that's what happens
if the keyword isn't used with MCell. Also, older data models will be
upgraded to include this new value with "WARNING" so they should produce
the same output as they had with the older version of MCell (before
dynamic geometry was available).

Note that there is a minor problem here because this keyword will be
written into all MDL files that CellBlender generates. This keyword was
not understood for older versions of MCell. We could run MCell first to
read back its version number to decide whether to output these kinds of
fields or not. We could also only write that field when a model actually
contains dynamic geometry objects. This remains to be decided. The same
is true for display of the value. It *might* only be displayed as an
option if dynamic geometry is detected. However, this can lead to some
confusion when fields appear and disappear. Besides, a user might want
to set that field first when starting a dynamic geometry project even
before creating any objects at all. It would be confusing for the
setting to not be available at that time. With those thoughts in mind,
the current implementation shows the field in the panel at all times.

The previous test suite code was still working "outside" of CellBlender.
It had used its own frame change handler and changed the exported MDL
(just as the IO_MESH version had done). A review of that test suite code
revealed that it was outputting both an MDL file and a binary frame file
for each frame of the dynamic geometry (for speed). This code could be
added to our dynamic geometry exporter as well - possibly as an option.

This commit added new tests that create Blender text objects and fill
them with the appropriate dynamic geometry script code. This is exactly
how the end user would do it. This is in keeping with the test suite's
role of trying to test CellBlender's response to actual user input by
invoking actual Blender mechanisms whenever possible.