Updated all of the dynamic geometry tests to use actual DG scripts.
Previous versions of the test suite had used special mechanisms to
generate dynamic geometry MDL and to display the dynamic geometry.
Those versions patched the MDL after it was exported, and they set up
their own frame change handler for visualization.

Over time, those mechanisms have been incorporated into CellBlender
itself. This commit updates the test suite to use CellBlender's own
built-in dynamic geometry scripting mechanisms. In general, that means
the test defines its own dynamic geometry code and inserts that code
into a Blender text object (text editor). Then the dynamic object flags
are set for the objects that use those scripts. At that point, the
normal CellBlender export and run capabilities can be used without any
need to "patch" the exported MDL to support dynamic geometry.

Note that these versions do not currently pass because the exported MDL
is NOT the same as was exported in previous versions. While they appear
(visually) to produce the same results, the MDL is exported differently,
and it is not byte-for-byte identical. Test passage is based on the SHA1
sum of the exported MDL being identical to the previous (supposed good)
exported MDL. There are two solutions:

  1. Work on the exporting code until it exports exactly the same MDL
  2. Judge that the new MDL is also correct and update the SHA1 sums

It is advised to follow the first approach to ensure that we get exactly
the same MDL. While this will take longer, it will be more likely to
uncover any problems.

Also note that the current exporting approach (using the data model)
does not yet apply any of the "Export Scripting" scripts to the MDL.
This functionality was implemented in the IO_Mesh exporting system and
has not yet been added to the data model exporting system. There have
been no test cases testing this functionality, but they should be added.