Modified mdl/data_model_to_mdl.py to export modular MDL files.
The main effort in this commit is the exporting of modular MDL
format from the data model. This will be helpful to compare the
new data model dynamic geometry exporting with the existing IO
mesh exporting (from Blender's properties). This should also be
supported in general since there may be cases where modular MDL
is the desired format.

To more easily compare upgrades, the "export_format" field was
added to the data model to preserve whether a model should be
exported as modular or unified.

This commit also added support for exporting point release
lists from the data model (an overlooked feature).

Note that exporting as modular MDL files from a data model
produces a slightly different physical file result than using
the IO mesh exporter. The IO Mesh exporter puts a number of
initialization values (ITERATIONS, TIME_STEP, PARTITIONS, etc)
into the main MDL file (Scene.main.mdl). In the data model,
these values are all considered part of the "initialization"
section. So it was most natural for the data model exporter
to export those fields in the Scene.initialization.mdl file
rather than in the main (Scene.main.mdl) file. They are,
however, exported in the same order when the modular files
are included. This was verified with a small Python script
that recursively assembles the modular files into a single
MDL file for comparison. Here's the script:

  def combine_mdl ( line_list, file_name ):
    try:
      f = open ( file_name, 'r' )
      lines = f.readlines();
      for line in lines:
        l = line.strip()
        if l.startswith('INCLUDE_FILE'):
          q = l.find('"')
          l = l[q+1:]
          q = l.find('"')
          l = l[0:q]
          combine_mdl ( line_list, l )
        else:
          line_list.append ( line )
      f.close()
    except FileNotFoundError as ioe:
        print ( "File not found error: " + str(e) )
    except Exception as e:
        print ( "Exception reading MDL: " + str(e) )
    except:
        print ( "Unknown Exception" )
    return ( line_list )

  ll = combine_mdl ( [], "Scene.main.mdl" )

  for l in ll:
    print ( l[0:-1] )

  print ( "Total of " + str(len(ll)) + " lines." )

That script can be run with the output directed into a
file to generate a unified MDL file from modular files.
This was done with an expanded version of the Organelle
demo run under Blender 2.78 (CBID: f2b1f4b4c0bdfca2...)
and 2.79 (CBID: ec40d735cbc729...). The resulting MDL
(as MCell should assemble it) differed only by:

  - Ordering of some parameters
  - Placement and amount of white space
  - Addition of the LARGE_MOLECULAR_DISPLACEMENT flag

This is not a complete test, and no testing has been
done with dynamic geometry at this point. That should
be the next step in preparing for the upcoming release.